requirements.txt
amqp==5.1.1
billiard==3.6.4.0
blinker==1.6.2
celery==5.2.7
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dnspython==2.3.0
docopt==0.6.2
email-validator==2.0.0.post2
Flask==2.2.5
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
gunicorn==20.1.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.1
pdfkit==1.0.0
prompt-toolkit==3.0.38
PyJWT==2.7.0
pytz==2023.3
requests==2.30.0
SQLAlchemy==2.0.13
urllib3==2.0.2
Werkzeug==2.2.3
